Move raft state to separate file
store.go is getting big.
